Move editor-specific utils to components/Editor/util/
