+++ ui/src/components/Editor/EditorControls/BetaDetailsDragPreview.tsx
@@ -1,5 +1,5 @@
 import { graphql, useFragment } from "react-relay";
-import { DragItemWithKind } from "util/dnd";
+import { DragItemWithKind } from "components/Editor/util/dnd";
 import { assertIsDefined, findNode } from "util/func";
 import BetaMoveListItem from "./BetaMoveListItem";
 import { BetaDetailsDragPreview_betaMoveNodeConnection$key } from "./__generated__/BetaDetailsDragPreview_betaMoveNodeConnection.graphql";
