+++ ui/src/components/Editor/EditorControls/BetaMoveListItem.tsx
@@ -11,7 +11,7 @@ import {
   ListItemSecondaryAction,
   ListItemText,
 } from "@mui/material";
-import { formatBodyPart, useBetaMoveColor } from "util/svg";
+import { formatBodyPart, useBetaMoveColor } from "components/Editor/util/svg";
 import { graphql, useFragment } from "react-relay";
 import { BetaMoveListItem_betaMoveNode$key } from "./__generated__/BetaMoveListItem_betaMoveNode.graphql";
 import { BetaMoveIconWrapped } from "../EditorSvg/BetaEditor/BetaMoveIcon";
