+++ ui/src/components/Editor/EditorSvg/HoldEditor/HoldEditor.tsx
@@ -2,10 +2,10 @@ import MutationErrorSnackbar from "components/common/MutationErrorSnackbar";
 import { useContext } from "react";
 import { useFragment } from "react-relay";
 import { graphql } from "relay-runtime";
-import { EditorSelectedBetaContext } from "util/context";
-import { DropHandler, getItemWithKind } from "util/dnd";
+import { EditorSelectedBetaContext } from "components/Editor/util/context";
+import { DropHandler, getItemWithKind } from "components/Editor/util/dnd";
 import { assertIsDefined } from "util/func";
-import { useHighlight } from "util/highlight";
+import { useHighlight } from "components/Editor/util/highlight";
 import useMutation from "util/useMutation";
 import HoldEditorDropZone from "./HoldEditorDropZone";
 import HoldOverlay from "./HoldOverlay";
