+++ ui/src/components/Editor/EditorSvg/common/Line.tsx
@@ -1,4 +1,4 @@
-import { OverlayPosition } from "util/svg";
+import { OverlayPosition } from "components/Editor/util/svg";
 
 interface Props extends React.SVGProps<SVGLineElement> {
   p1: OverlayPosition;
